move tweet cli to separate file
