+++ busy_beaver/apps/retweeter/blueprint.py
@@ -2,4 +2,4 @@ from flask import blueprints
 
 twitter_bp = blueprints.Blueprint("twitter", __name__)
 
-from . import task  # noqa isort:skip
+from . import cli  # noqa isort:skip
