+++ busy_beaver/apps/retweeter/task.py
@@ -1,27 +1,14 @@
 from datetime import timedelta
 import logging
 
-import click
-
-from .blueprint import twitter_bp
 from busy_beaver.clients import twitter
 from busy_beaver.common.wrappers import KeyValueStoreClient, SlackClient
-from busy_beaver.config import TWITTER_USERNAME
 from busy_beaver.models import SlackInstallation
 from busy_beaver.toolbox import utc_now_minus
 
 LAST_TWEET_KEY = "last_posted_tweet_id"
 logger = logging.getLogger(__name__)
 kv_store = KeyValueStoreClient()
-
-
-@click.option("--channel_name", required=True, help="Slack channel")
-@click.option("--workspace", required=True, help="Slack workspace ID")
-@twitter_bp.cli.command("poll_twitter", help="Find new tweets to post to Slack")
-def poll_twitter(channel_name: str, workspace: str):
-    # TODO add logging and times
-    installation = SlackInstallation.query.filter_by(workspace_id=workspace).first()
-    fetch_tweets_post_to_slack(installation.id, channel_name, username=TWITTER_USERNAME)
 
 
 def fetch_tweets_post_to_slack(installation_id, channel_name, username):
