+++ tests/apps/retweeter/task_test.py
@@ -2,11 +2,8 @@ from datetime import timedelta
 
 import pytest
 
-from busy_beaver.apps.retweeter.task import (
-    LAST_TWEET_KEY,
-    fetch_tweets_post_to_slack,
-    poll_twitter,
-)
+from busy_beaver.apps.retweeter.cli import poll_twitter
+from busy_beaver.apps.retweeter.task import LAST_TWEET_KEY, fetch_tweets_post_to_slack
 from busy_beaver.toolbox import utc_now_minus
 from tests._utilities import FakeSlackClient
